+++ src/app/collective-rewards/rewards/backers/index.ts
@@ -1,3 +1,3 @@
 export * from './context'
 export * from './hooks'
-export * from './actions/backerStakingHistoryAction'
+export * from './actions/fetchBackerStakingHistory'
