+++ src/shared/components/ApolloClient/makeClient.ts
@@ -10,6 +10,7 @@ function makeClient(uri: string) {
     // you can disable result caching here if you want to
     // (this does not work if you are rendering your page with `export const dynamic = "force-static"`)
     fetchOptions: {
+      cache: 'no-store',
       // you can pass additional options that should be passed to `fetch` here,
       // e.g. Next.js-related `fetch` options regarding caching and revalidation
       // see https://nextjs.org/docs/app/api-reference/functions/fetch#fetchurl-options
